+++ examples/axum/src/db/migrations/20240530135613_create_tables_categories_and_products.down.sql
@@ -1,0 +1,4 @@
+-- Add down migration script here
+DROP TABLE IF EXISTS inventory.products;
+
+DROP TABLE IF EXISTS inventory.categories;